==> Code/02-Population.ipynb <==
import pandas as pd
import re

**· 2011 London LSOA Population**

# Load CSV file
file_path_lsoa11 = 'Data/Raw/Population/LSOA_11.csv'
df_lsoa11= pd.read_csv(file_path_lsoa11)

# Specify the list of borough names to retain
boroughs = [
    'City of London', 'Barking and Dagenham', 'Barnet', 'Bexley', 'Brent', 'Bromley', 
    'Camden', 'Croydon', 'Ealing', 'Enfield', 'Greenwich', 'Hackney', 
    'Hammersmith and Fulham', 'Haringey', 'Harrow', 'Havering', 'Hillingdon', 
    'Hounslow', 'Islington', 'Kensington and Chelsea', 'Kingston upon Thames', 
    'Lambeth', 'Lewisham', 'Merton', 'Newham', 'Redbridge', 'Richmond upon Thames', 
    'Southwark', 'Sutton', 'Tower Hamlets', 'Waltham Forest', 'Wandsworth', 'Westminster'
]

# Check if LSOA Name contains the exact borough name
pattern = '|'.join([f'\\b{borough}\\b' for borough in boroughs])

# Filter the data
filtered_df_lsoa11 = df_lsoa11[df_lsoa11['LSOA Name'].str.contains(pattern, regex=True)]

# Save the filtered DataFrame to a new CSV file
filtered_df_lsoa11.to_csv('Data/Revised/Population/LSOA_2011.csv', index=False)

**· 2011 London MSOA Population**

# Load CSV file
file_path_msoa11 = 'Data/Raw/Population/MSOA_11.csv'
df_msoa11= pd.read_csv(file_path_msoa11)

# Specify the list of borough names to retain
boroughs = [
    'City of London', 'Barking and Dagenham', 'Barnet', 'Bexley', 'Brent', 'Bromley', 
    'Camden', 'Croydon', 'Ealing', 'Enfield', 'Greenwich', 'Hackney', 
    'Hammersmith and Fulham', 'Haringey', 'Harrow', 'Havering', 'Hillingdon', 
    'Hounslow', 'Islington', 'Kensington and Chelsea', 'Kingston upon Thames', 
    'Lambeth', 'Lewisham', 'Merton', 'Newham', 'Redbridge', 'Richmond upon Thames', 
    'Southwark', 'Sutton', 'Tower Hamlets', 'Waltham Forest', 'Wandsworth', 'Westminster'
]

# Check if MSOA Name contains the exact borough name
pattern = '|'.join([f'\\b{borough}\\b' for borough in boroughs])

# Filter the data
filtered_df_msoa11 = df_msoa11[df_msoa11['MSOA Name'].str.contains(pattern, regex=True)]

# Save the filtered DataFrame to a new CSV file
filtered_df_msoa11.to_csv('Data/Revised/Population/MSOA_2011.csv', index=False)

**· 2021 London MSOA Population**

# Load CSV file
file_path_msoa21 = 'Data/Raw/Population/MSOA_21.csv'
df_msoa21= pd.read_csv(file_path_msoa21)

# Specify the list of borough names to retain
boroughs = [
    'City of London', 'Barking and Dagenham', 'Barnet', 'Bexley', 'Brent', 'Bromley', 
    'Camden', 'Croydon', 'Ealing', 'Enfield', 'Greenwich', 'Hackney', 
    'Hammersmith and Fulham', 'Haringey', 'Harrow', 'Havering', 'Hillingdon', 
    'Hounslow', 'Islington', 'Kensington and Chelsea', 'Kingston upon Thames', 
    'Lambeth', 'Lewisham', 'Merton', 'Newham', 'Redbridge', 'Richmond upon Thames', 
    'Southwark', 'Sutton', 'Tower Hamlets', 'Waltham Forest', 'Wandsworth', 'Westminster'
]

# Check if MSOA Name contains the exact borough name
pattern = '|'.join([f'\\b{borough}\\b' for borough in boroughs])

# Filter the data
filtered_df_msoa21 = df_msoa21[df_msoa21['MSOA 2021 Name'].str.contains(pattern, regex=True)]

# Save the filtered DataFrame to a new CSV file
filtered_df_msoa21.to_csv('Data/Revised/Population/MSOA_2021.csv', index=False)

**· 2021 London LSOA Population**

# Load CSV file
file_path_msoa21 = 'Data/Raw/Population/LSOA_21.csv'
df_msoa21= pd.read_csv(file_path_msoa21)

# Specify the list of borough names to retain
boroughs = [
    'City of London', 'Barking and Dagenham', 'Barnet', 'Bexley', 'Brent', 'Bromley', 
    'Camden', 'Croydon', 'Ealing', 'Enfield', 'Greenwich', 'Hackney', 
    'Hammersmith and Fulham', 'Haringey', 'Harrow', 'Havering', 'Hillingdon', 
    'Hounslow', 'Islington', 'Kensington and Chelsea', 'Kingston upon Thames', 
    'Lambeth', 'Lewisham', 'Merton', 'Newham', 'Redbridge', 'Richmond upon Thames', 
    'Southwark', 'Sutton', 'Tower Hamlets', 'Waltham Forest', 'Wandsworth', 'Westminster'
]

# Check if MSOA Name contains the exact borough name
pattern = '|'.join([f'\\b{borough}\\b' for borough in boroughs])

# Filter the data
filtered_df_msoa21 = df_msoa21[df_msoa21['LSOA 2021 Name'].str.contains(pattern, regex=True)]

# Save the filtered DataFrame to a new CSV file
filtered_df_msoa21.to_csv('Data/Revised/Population/LSOA_2021.csv', index=False)
