==> Code/03-Lookup.ipynb <==
import pandas as pd

**· 2011 LSOA - 2011 MSOA**

# Load CSV file
file_path_11 = 'Data/Raw/Geography/Lookup/11-11.csv'
df_11 = pd.read_csv(file_path_11)

# Specify the list of area names to retain
areas = [
    "City of London", "Barking and Dagenham", "Barnet", "Bexley", "Brent", "Bromley", "Camden", "Croydon",
    "Ealing", "Enfield", "Greenwich", "Hackney", "Hammersmith and Fulham", "Haringey", "Harrow", "Havering",
    "Hillingdon", "Hounslow", "Islington", "Kensington and Chelsea", "Kingston upon Thames", "Lambeth", 
    "Lewisham", "Merton", "Newham", "Redbridge", "Richmond upon Thames", "Southwark", "Sutton", 
    "Tower Hamlets", "Waltham Forest", "Wandsworth", "Westminster"
]

# Create filter conditions using regex to ensure exact matches
pattern = r'\b(?:' + '|'.join(areas) + r')\b'

# Filter the DataFrame
filtered_df_11 = df_11[df_11['MSOA11NM'].str.contains(pattern, na=False, regex=True) | df_11['LSOA11NM'].str.contains(pattern, na=False, regex=True)]

# Remove rows with duplicate 'LSOA11CD' or 'LSOA11NM'
cleaned_df_11 = filtered_df_11.drop_duplicates(subset=['LSOA11CD', 'LSOA11NM'], keep='first')

# Save the filtered DataFrame to a new CSV file
cleaned_df_11.to_csv('Data/Revised/Geography/Lookup/11-11-MSOA.csv', index=False)

**· 2011 MSOA - 2021 MSOA**

# Load CSV file
file_path_21 = 'Data/Raw/Geography/Lookup/11-21.csv'
df_21 = pd.read_csv(file_path_21)

# Specify the list of area names to retain
areas = [
    "City of London", "Barking and Dagenham", "Barnet", "Bexley", "Brent", "Bromley", "Camden", "Croydon",
    "Ealing", "Enfield", "Greenwich", "Hackney", "Hammersmith and Fulham", "Haringey", "Harrow", "Havering",
    "Hillingdon", "Hounslow", "Islington", "Kensington and Chelsea", "Kingston upon Thames", "Lambeth", 
    "Lewisham", "Merton", "Newham", "Redbridge", "Richmond upon Thames", "Southwark", "Sutton", 
    "Tower Hamlets", "Waltham Forest", "Wandsworth", "Westminster"
]

# Create filter conditions using regex to ensure exact matches
pattern = r'\b(?:' + '|'.join(areas) + r')\b'

# Filter the DataFrame
filtered_df_21 = df_21[df_21['MSOA11NM'].str.contains(pattern, na=False, regex=True) | df_21['MSOA21NM'].str.contains(pattern, na=False, regex=True)]

# Remove rows with duplicate 'MSOA21CD' or 'MSOA21NM'
cleaned_df_21 = filtered_df_21.drop_duplicates(subset=['MSOA21CD', 'MSOA21NM'], keep='first')

# Save the filtered DataFrame to a new CSV file
cleaned_df_21.to_csv('Data/Revised/Geography/Lookup/11-21-MSOA.csv', index=False)

**· 2021 MSOA - 2021 LSOA**

# Load CSV file
file_path_ml = 'Data/Raw/Geography/Lookup/MSOA-LSOA.csv'
df_ml = pd.read_csv(file_path_ml)

# Specify the list of area names to retain
areas = [
    "City of London", "Barking and Dagenham", "Barnet", "Bexley", "Brent", "Bromley", "Camden", "Croydon",
    "Ealing", "Enfield", "Greenwich", "Hackney", "Hammersmith and Fulham", "Haringey", "Harrow", "Havering",
    "Hillingdon", "Hounslow", "Islington", "Kensington and Chelsea", "Kingston upon Thames", "Lambeth", 
    "Lewisham", "Merton", "Newham", "Redbridge", "Richmond upon Thames", "Southwark", "Sutton", 
    "Tower Hamlets", "Waltham Forest", "Wandsworth", "Westminster"
]

# Create filter conditions using regex to ensure exact matches
pattern = r'\b(?:' + '|'.join(areas) + r')\b'

# Filter the DataFrame
filtered_df_ml = df_ml[df_ml['MSOA21NM'].str.contains(pattern, na=False, regex=True) | df_ml['LSOA21NM'].str.contains(pattern, na=False, regex=True)]

# Remove rows with duplicate 'MSOA21CD' or 'MSOA21NM'
cleaned_df_ml = filtered_df_ml.drop_duplicates(subset=['LSOA21CD', 'LSOA21NM'], keep='first')

# Save the filtered DataFrame to a new CSV file
cleaned_df_ml.to_csv('Data/Revised/Geography/Lookup/11-21-ML.csv', index=False)
