==> Code/04-ODMG.ipynb <==
import pandas as pd

**· Origin-Destination Migration data**

# Load CSV file
file_path = 'Data/Raw/Migration/ODMG.csv'
df = pd.read_csv(file_path)

# Remove rows with invalid values
filtered_df = df[
    ~df['Migrant MSOA one year ago code'].isin(['-8', '999999999']) &
    ~df['Migrant MSOA one year ago label'].isin(['Does not apply', 'Person with migrant origin outside the UK'])
]

# Define the list of keywords for the areas to include
areas = [
    "City of London", "Barking and Dagenham", "Barnet", "Bexley", "Brent", "Bromley", "Camden", "Croydon",
    "Ealing", "Enfield", "Greenwich", "Hackney", "Hammersmith and Fulham", "Haringey", "Harrow", "Havering",
    "Hillingdon", "Hounslow", "Islington", "Kensington and Chelsea", "Kingston upon Thames", "Lambeth", 
    "Lewisham", "Merton", "Newham", "Redbridge", "Richmond upon Thames", "Southwark", "Sutton", 
    "Tower Hamlets", "Waltham Forest", "Wandsworth", "Westminster"
]

# Filter rows by checking for exact matches with the keywords in the relevant columns
filtered_df_strict = filtered_df[
    filtered_df['Migrant MSOA one year ago label'].apply(lambda x: any(area in x for area in areas)) &
    filtered_df['Middle layer Super Output Areas label'].apply(lambda x: any(area in x for area in areas))
]

# Save the filtered DataFrame to a new CSV file
filtered_df_strict.to_csv('Data/Revised/Migration/ODMG.csv', index=False)
