==> Code/08-IMD2021.ipynb <==
import pandas as pd

# Load the data files
file1_path = "Data/Revised/Geography/Lookup/11-21-MSOA.csv"
file2_path = "Data/Revised/Geography/MSOA/2011.csv"
file3_path = "Data/Revised/Geography/MSOA/2021.csv"

# Reading the CSV files
file1 = pd.read_csv(file1_path)
file2 = pd.read_csv(file2_path)
file3 = pd.read_csv(file3_path)

# Merging the first file with 2011 data based on 'MSOA11CD'
merged_data_2011 = file1.merge(file2[['MSOA11CD', 'area_2011']], on='MSOA11CD', how='left')

# Merging the result with 2021 data based on 'msoa21cd'
merged_final_data = merged_data_2011.merge(file3[['msoa21cd', 'area_2021']], left_on='MSOA21CD', right_on='msoa21cd', how='left')

# Dropping the redundant 'msoa21cd' column after merging
merged_final_data.drop(columns=['msoa21cd'], inplace=True)

# Calculating the ratio of area_2011 to area_2021 and adding it as a new column 'Ratio'
merged_final_data['Ratio'] = (merged_final_data['area_2021'] / merged_final_data['area_2011']).round(2)

# Display the updated dataframe
print(merged_final_data)

# Load the IMD 2019 data
imd_2019_path = "Data/Revised/Migration/IMD_2019.csv"
imd_2019_data = pd.read_csv(imd_2019_path)

# Merging the existing data with the IMD 2019 data based on 'MSOA11CD'
merged_with_score = merged_final_data.merge(imd_2019_data[['MSOA11CD', 'Score']], on='MSOA11CD', how='left')

# Renaming the 'Score' column to 'Score_2019'
merged_with_score.rename(columns={'Score': 'Score_2019'}, inplace=True)

# Calculating the ratio of area_2021 to area_2011
merged_with_score['Ratio'] = (merged_with_score['area_2021'] / merged_with_score['area_2011']).round(2)

# Calculating the product of 'Ratio' and 'Score_2019', rounded to two decimal places
merged_with_score['Score'] = (merged_with_score['Ratio'] * merged_with_score['Score_2019']).round(2)

# Calculating Decile rankings based on the 'Score' column, with 1 being the highest score
merged_with_score['Decile'] = pd.qcut(merged_with_score['Score'], 10, labels=range(10, 0, -1))

# Converting the Decile values to integers
merged_with_score['Decile'] = merged_with_score['Decile'].astype(int)

# Display the updated dataframe
print(merged_with_score)

# Selecting only the specified columns
final_data = merged_with_score[['MSOA21CD', 'MSOA21NM', 'Score', 'Decile']]

# Exporting the selected data to a new CSV file
final_data.to_csv('Data/Revised/Migration/IMD_2021.csv', index=False)
