==> Code/10-Final Dataset.ipynb <==
import pandas as pd

# Load the CSV files into DataFrames
health_2011 = pd.read_csv('Data/Revised/Health/Health_2011.csv')
health_2021 = pd.read_csv('Data/Revised/Health/Health_2021.csv')
migration = pd.read_csv('Data/Revised/Migration/Migration.csv')

# Merge Health_2011 with Migration based on the corresponding codes
merged_data = migration.merge(
    health_2011[['MSOA21CD', 'Decile_Before']], 
    left_on='Code_2020', 
    right_on='MSOA21CD', 
    how='left'
)

# Rename the column '2011' to 'Health_Before'
merged_data.rename(columns={'Decile_Before': 'Health_Before'}, inplace=True)

# Drop the extra MSOA21CD column from the merge
merged_data.drop(columns=['MSOA21CD'], inplace=True)

# Merge Health_2021 with the previously merged data based on the corresponding codes
merged_data = merged_data.merge(
    health_2021[['MSOA21CD', 'Decile_After']], 
    left_on='Code_2021', 
    right_on='MSOA21CD', 
    how='left'
)

# Rename the column '2021' to 'Health_After'
merged_data.rename(columns={'Decile_After': 'Health_After'}, inplace=True)

# Drop the extra MSOA21CD column from the merge
merged_data.drop(columns=['MSOA21CD'], inplace=True)

# Calculate the difference between Health_After and Health_Before to form the new column Health_Change
merged_data['Health_Change'] = merged_data['Health_After'] - merged_data['Health_Before']

# Move the 'Count' column to the last position
count_column = merged_data.pop('Count')
merged_data['Count'] = count_column

# Export the modified DataFrame to a new CSV file
output_path = "Data/Final/Dataset.csv"
merged_data.to_csv(output_path, index=False)

# Load the CSV file
file_path = "Data/Final/Dataset.csv"
data = pd.read_csv(file_path)

# Calculate new columns 'IMD' and 'Health'
data['IMD_Total'] = data['IMD_Change'] * data['Count']
data['Health_Total'] = data['Health_Change'] * data['Count']

# Remove only rows with missing values
data = data.dropna()

# Export the updated dataframe to a new CSV file
export_path = "Data/Final/Dataset.csv"
data.to_csv(export_path, index=False)
